#495 reorg step 5 aft_face ideas to fiscal jsons etl
