+++ src/f11_world/test/test_world_etl05_04_fiscal_csvs_to_fiscal_json.py
@@ -1,35 +1,8 @@
 from src.f00_instrument.file import create_path, save_file, open_file
-from src.f01_road.finance_tran import bridge_str
-from src.f03_chrono.chrono import (
-    c400_number_str,
-    yr1_jan1_offset_str,
-    monthday_distortion_str,
-    timeline_title_str,
-    timelineunit_shop,
-    timeline_config_shop,
-)
-from src.f04_gift.atom_config import (
-    acct_name_str,
-    face_name_str,
-    fiscal_title_str,
-    owner_name_str,
-    fund_coin_str,
-    penny_str,
-    respect_bit_str,
-)
+from src.f03_chrono.chrono import timelineunit_shop, timeline_config_shop
 from src.f07_fiscal.fiscal import (
     fiscalunit_shop,
     get_from_json as fiscalunit_get_from_json,
-)
-from src.f07_fiscal.fiscal_config import (
-    get_fiscal_config_args,
-    fiscalunit_str,
-    fiscal_deal_episode_str,
-    fiscal_cashbook_str,
-    fiscal_timeline_hour_str,
-    fiscal_timeline_month_str,
-    fiscal_timeline_weekday_str,
-    present_time_str,
 )
 from src.f10_etl.fiscal_etl_tool import (
     FiscalPrimeObjsRef,
@@ -37,7 +10,6 @@ from src.f10_etl.fiscal_etl_tool import (
 )
 from src.f11_world.world import worldunit_shop
 from src.f11_world.examples.world_env import get_test_worlds_dir, env_dir_setup_cleanup
-from copy import copy as copy_copy
 from os.path import exists as os_path_exists
